PR fixes and mode2D fix for modular scenarios.
