+++ src/scenic/core/simulators.py
@@ -809,11 +809,11 @@ class Simulation(abc.ABC):
         The definition of 'state' is up to the simulator; the 'state' is simply saved
         at each time step to define the 'trajectory' of the simulation. Changing this
         method is however discouraged, unless one is adding additional attributes to
-        the returned `SimulationState` object.
+        the returned ``SimulationState`` object.
 
-        The default implementation returns a custom `SimulationState` object, which is a tuple
+        The default implementation returns a custom ``SimulationState`` object, which is a tuple
         of positions of all objects (for backwards compatibility) and also has two attributes:
-        `positions` and `orientations`, which are themselves tuples of the positions and
+        ``positions`` and ``orientations``, which are themselves tuples of the positions and
         orientations of all objects.
         """
