+++ tests/domains/driving/test_driving.py
@@ -7,7 +7,9 @@ import pytest
 from scenic.core.distributions import RejectionException
 from scenic.core.errors import InvalidScenarioError
 from scenic.core.geometry import TriangulationError
+from scenic.domains.driving.openscenario import toOpenScenario
 from scenic.domains.driving.roads import Network
+from scenic.simulators.newtonian import NewtonianSimulator
 from tests.utils import compileScenic, pickle_test, sampleEgo, sampleScene, tryPickling
 
 # Suppress all warnings from OpenDRIVE parser
@@ -229,3 +231,69 @@ def test_invalid_road_scenario(cached_maps):
             param foo = ego.lane
             """,
         )
+
+
+def test_xosc_export(getAssetPath):
+    simulator = NewtonianSimulator("Town01")
+    code = f"""
+        param render = False
+        model scenic.simulators.newtonian.driving_model
+
+        ego = new Car with behavior FollowLaneBehavior()
+
+        immobileCar = new Car visible
+
+        behavior Walk():
+            take SetWalkingSpeedAction(1)
+        
+        new Pedestrian behind ego by 5,
+            with regionContainedIn None,
+            with behavior Walk()
+
+        terminate after 5 seconds
+    """
+
+    scenario = compileScenic(
+        code, mode2D=True, params={"map": getAssetPath("maps/CARLA/Town01.xodr")}
+    )
+    scene, _ = scenario.generate()
+    simulation = simulator.simulate(scene)
+    toOpenScenario(simulation, scenario, scene)
+
+
+def test_xosc_export_dynamic_objects(getAssetPath):
+    simulator = NewtonianSimulator("Town01")
+    code = f"""
+        param render = False
+        model scenic.simulators.newtonian.driving_model
+
+        scenario SpawnCar():
+            setup:
+                new Car at 0@0
+
+        scenario Main():
+            setup:
+                ego = new Car with behavior FollowLaneBehavior()
+
+                immobileCar = new Car visible
+
+                behavior Walk():
+                    take SetWalkingSpeedAction(1)
+                
+                new Pedestrian behind ego by 5,
+                    with regionContainedIn None,
+                    with behavior Walk()
+
+                terminate after 5 seconds
+
+            compose:
+                wait for 1 seconds
+                do SpawnCar()
+    """
+    scenario = compileScenic(
+        code, mode2D=True, params={"map": getAssetPath("maps/CARLA/Town01.xodr")}
+    )
+    scene, _ = scenario.generate()
+    simulation = simulator.simulate(scene)
+    with pytest.raises(RuntimeError, match="(.*)dynamically created objects(.*)"):
+        toOpenScenario(simulation, scenario, scene)
